Spotbugs excludes in code.
