+++ all/pom.xml
@@ -1117,6 +1117,10 @@
             <artifactId>helidon-metadata-hson</artifactId>
         </dependency>
         <dependency>
+            <groupId>io.helidon.metadata</groupId>
+            <artifactId>helidon-metadata-compile</artifactId>
+        </dependency>
+        <dependency>
             <groupId>io.helidon.integrations.oci.sdk</groupId>
             <artifactId>helidon-integrations-oci-sdk-processor</artifactId>
         </dependency>
