+++ bom/pom.xml
@@ -1468,6 +1468,11 @@
                 <artifactId>helidon-metadata-hson</artifactId>
                 <version>${helidon.version}</version>
             </dependency>
+            <dependency>
+                <groupId>io.helidon.metadata</groupId>
+                <artifactId>helidon-metadata-compile</artifactId>
+                <version>${helidon.version}</version>
+            </dependency>
 
             <!-- Injection Integrations related -->
             <dependency>
