--- config/config/etc/spotbugs/exclude.xml
@@ -1,107 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2019, 2022 Oracle and/or its affiliates.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-
-<FindBugsFilter
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xmlns="https://github.com/spotbugs/filter/3.0.0"
-        xsi:schemaLocation="https://github.com/spotbugs/filter/3.0.0 https://raw.githubusercontent.com/spotbugs/spotbugs/3.1.0/spotbugs/etc/findbugsfilter.xsd">
-
-    <Match>
-        <!-- Hash is used only to see if a file has changed -->
-        <Class name="io.helidon.config.FileSourceHelper"/>
-        <Bug pattern="UNSAFE_HASH_EQUALS"/>
-    </Match>
-    <Match>
-        <!-- Hash is used only to see if a file has changed -->
-        <Class name="io.helidon.config.FileSourceHelper"/>
-        <Bug pattern="WEAK_MESSAGE_DIGEST_MD5"/>
-    </Match>
-    <Match>
-        <!-- Path from config/code, not openly from user -->
-        <Class name="io.helidon.config.ConfigMappers"/>
-        <Bug pattern="PATH_TRAVERSAL_IN"/>
-    </Match>
-    <Match>
-        <!-- Path from config/code, not openly from user -->
-        <Class name="io.helidon.config.ConfigSources"/>
-        <Bug pattern="PATH_TRAVERSAL_IN"/>
-    </Match>
-    <Match>
-        <!-- Path from config/code, not openly from user -->
-        <Class name="io.helidon.config.OverrideSources"/>
-        <Bug pattern="PATH_TRAVERSAL_IN"/>
-    </Match>
-    <Match>
-        <!-- Path from config/code/system property, not openly from user -->
-        <Class name="io.helidon.config.MetaConfigFinder"/>
-        <Bug pattern="PATH_TRAVERSAL_IN"/>
-    </Match>
-    <Match>
-        <!-- Path from config/code, not openly from user -->
-        <Class name="io.helidon.config.ClasspathConfigSource"/>
-        <Bug pattern="URLCONNECTION_SSRF_FD"/>
-    </Match>
-    <Match>
-        <!-- Path from config/code, not openly from user -->
-        <Class name="io.helidon.config.ClasspathOverrideSource"/>
-        <Bug pattern="URLCONNECTION_SSRF_FD"/>
-    </Match>
-    <Match>
-        <!-- Path from config/code, not openly from user -->
-        <Class name="io.helidon.config.UrlConfigSource"/>
-        <Bug pattern="URLCONNECTION_SSRF_FD"/>
-    </Match>
-    <Match>
-        <!-- Path from config/code, not openly from user -->
-        <Class name="io.helidon.config.UrlHelper"/>
-        <Bug pattern="URLCONNECTION_SSRF_FD"/>
-    </Match>
-    <Match>
-        <!-- Path from config/code, not openly from user -->
-        <Class name="io.helidon.config.UrlOverrideSource"/>
-        <Bug pattern="URLCONNECTION_SSRF_FD"/>
-    </Match>
-    <Match>
-        <!-- False positive. See https://github.com/spotbugs/spotbugs/issues/756 -->
-        <Class name="io.helidon.config.PropertiesConfigParser"/>
-        <Method name="parse"/>
-        <Bug pattern="RCN_REDUNDANT_NULLCHECK_WOULD_HAVE_BEEN_A_NPE"/>
-    </Match>
-    <Match>
-        <!-- inheritance of the same interface from common/config -->
-        <Class name="io.helidon.config.Config"/>
-        <Bug pattern="NM_SAME_SIMPLE_NAME_AS_INTERFACE"/>
-    </Match>
-    <Match>
-        <!-- inheritance of the same interface from common/config -->
-        <Class name="io.helidon.config.Config$Key"/>
-        <Bug pattern="NM_SAME_SIMPLE_NAME_AS_INTERFACE"/>
-    </Match>
-    <Match>
-        <!-- inheritance of the same interface from common/config -->
-        <Class name="io.helidon.config.ConfigValue"/>
-        <Bug pattern="NM_SAME_SIMPLE_NAME_AS_INTERFACE"/>
-    </Match>
-    <Match>
-        <!-- inheritance of the same interface from common/config -->
-        <Class name="io.helidon.config.ConfigException"/>
-        <Bug pattern="NM_SAME_SIMPLE_NAME_AS_SUPERCLASS"/>
-    </Match>
-
-</FindBugsFilter>
