+++ config/config/src/main/java/io/helidon/config/PropertiesConfigParser.java
@@ -26,6 +26,7 @@ import io.helidon.common.media.type.MediaTypes;
 import io.helidon.config.spi.ConfigNode;
 import io.helidon.config.spi.ConfigParser;
 import io.helidon.config.spi.ConfigParserException;
+import io.helidon.metadata.compile.Spotbugs;
 
 /**
  * {@link ConfigParser} implementation that parses Java Properties content.
