+++ config/config/src/main/java/module-info.java
@@ -28,6 +28,7 @@ import io.helidon.common.features.api.HelidonFlavor;
 )
 module io.helidon.config {
     requires static io.helidon.common.features.api;
+    requires static io.helidon.metadata.compile;
 
     requires static io.helidon.service.registry;
     requires transitive io.helidon.common.config;
