--- config/metadata/codegen/src/main/java/io/helidon/config/metadata/codegen/TypeHandlerResult.java
@@ -1,31 +1,0 @@
-/*
- * Copyright (c) 2023, 2024 Oracle and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.helidon.config.metadata.codegen;
-
-import io.helidon.common.types.TypeName;
-
-/**
- * Result of annotation processing.
- *
- * @param targetType     type that is configured (result of the builder, runtime type of a prototype)
- * @param moduleName     module of the type
- * @param configuredType collected configuration metadata
- */
-record TypeHandlerResult(TypeName targetType,
-                         String moduleName,
-                         ConfiguredType configuredType) {
-}
