--- config/metadata/codegen/src/main/java/module-info.java
@@ -1,30 +1,0 @@
-/*
- * Copyright (c) 2024 Oracle and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Codegen for Helidon Config Metadata.
- *
- * @see io.helidon.config.metadata.codegen
- */
-module io.helidon.config.metadata.codegen {
-    requires io.helidon.codegen;
-    requires io.helidon.metadata.hson;
-
-    exports io.helidon.config.metadata.codegen;
-
-    provides io.helidon.codegen.spi.CodegenExtensionProvider
-            with io.helidon.config.metadata.codegen.ConfigMetadataCodegenProvider;
-}
