+++ metadata/codegen/src/main/java/io/helidon/metadata/codegen/spotbugs/SpotbugsTypes.java
@@ -1,0 +1,10 @@
+package io.helidon.metadata.codegen.spotbugs;
+
+import io.helidon.common.types.TypeName;
+
+final class SpotbugsTypes {
+    static final TypeName EXCLUDE = TypeName.create("io.helidon.metadata.compile.Spotbugs.Exclude");
+
+    private SpotbugsTypes() {
+    }
+}
