+++ metadata/compile/src/main/java/module-info.java
@@ -1,0 +1,3 @@
+module io.helidon.metadata.compile {
+    exports io.helidon.metadata.compile;
+}
